switched to floats

+++ inc/kernels.cuh
@@ -10,7 +10,7 @@ __global__
 void init_random_points(Point* points);
 
 __global__
-void init_points_in_circle(Point* points, double radius, double centre_x, double centre_y);
+void init_points_in_circle(Point* points, float radius, float centre_x, float centre_y);
 
 __global__
 void update_pos(Point* points);
